requirements.txt
# python-open3d
open3d-python
tqdm
numpy
pillow
opencv-python
pandas